perf: Server-Side Data Fetching in App Router: SettingsLayout / AdminLayout
