+++ apps/web/app/future/settings/(admin)/admin/page.tsx
@@ -10,4 +10,4 @@ export const generateMetadata = async () =>
     () => "admin_description"
   );
 
-export default WithLayout({ getLayout, Page: LegacyPage })<"P">;
+export default WithLayout({ getServerLayout: getLayout, Page: LegacyPage })<"P">;
