+++ apps/web/app/future/settings/(admin)/layout.tsx
@@ -2,4 +2,4 @@ import { WithLayout } from "app/layoutHOC";
 
 import { getLayout } from "@components/auth/layouts/AdminLayoutAppDir";
 
-export default WithLayout({ getLayout })<"L">;
+export default WithLayout({ getServerLayout: getLayout })<"L">;
