+++ apps/web/app/future/settings/(settings)/layout.tsx
@@ -2,4 +2,4 @@ import { WithLayout } from "app/layoutHOC";
 
 import { getLayout } from "@calcom/features/settings/appDir/SettingsLayoutAppDir";
 
-export default WithLayout({ getLayout: getLayout })<"L">;
+export default WithLayout({ getServerLayout: getLayout })<"L">;
