+++ apps/web/app/future/settings/(settings)/organizations/appearance/layout.tsx
@@ -2,4 +2,4 @@ import { WithLayout } from "app/layoutHOC";
 
 import { getLayout } from "@calcom/features/settings/appDir/SettingsLayoutAppDir";
 
-export default WithLayout({ getLayout });
+export default WithLayout({ getServerLayout: getLayout });
