+++ apps/web/app/future/settings/(settings)/teams/layout.tsx
@@ -2,4 +2,4 @@ import { WithLayout } from "app/layoutHOC";
 
 import { getLayout } from "@calcom/features/settings/appDir/SettingsLayoutAppDir";
 
-export default WithLayout({ getLayout })<"L">;
+export default WithLayout({ getServerLayout: getLayout })<"L">;
