+++ apps/web/app/layoutHOC.tsx
@@ -7,8 +7,10 @@ import { buildLegacyCtx } from "@lib/buildLegacyCtx";
 import PageWrapper from "@components/PageWrapperAppDir";
 
 type WithLayoutParams<T extends Record<string, any>> = {
-  getLayout: ((page: React.ReactElement) => React.ReactNode) | null;
+  getLayout?: ((page: React.ReactElement) => React.ReactNode) | null;
+  getServerLayout?: (page: React.ReactElement) => Promise<React.ReactNode | null>;
   Page?: (props: T) => React.ReactElement | null;
+  ServerPage?: (props: T) => Promise<React.ReactElement> | null;
   getData?: (arg: GetServerSidePropsContext) => Promise<T | undefined>;
   isBookingPage?: boolean;
   requiresLicense?: boolean;
@@ -16,7 +18,9 @@ type WithLayoutParams<T extends Record<string, any>> = {
 
 export function WithLayout<T extends Record<string, any>>({
   getLayout,
+  getServerLayout,
   getData,
+  ServerPage,
   Page,
   isBookingPage,
   requiresLicense,
@@ -30,7 +34,16 @@ export function WithLayout<T extends Record<string, any>>({
       props = (await getData(buildLegacyCtx(h, cookies(), p.params, p.searchParams))) ?? ({} as T);
     }
 
-    const children = "children" in p ? p.children : null;
+    // `p.children` exists only for layout.tsx files
+    const childrenFromLayoutFile = "children" in p ? p.children : null;
+    const page = ServerPage ? (
+      await ServerPage({ ...props, ...p })
+    ) : Page ? (
+      <Page {...props} />
+    ) : (
+      childrenFromLayoutFile
+    );
+    const pageWithServerLayout = page ? (getServerLayout ? await getServerLayout(page) : page) : null;
 
     return (
       <PageWrapper
@@ -40,7 +53,7 @@ export function WithLayout<T extends Record<string, any>>({
         themeBasis={null}
         isBookingPage={isBookingPage || !!(Page && "isBookingPage" in Page && Page.isBookingPage)}
         {...props}>
-        {Page ? <Page {...props} /> : children}
+        {pageWithServerLayout}
       </PageWrapper>
     );
   };
